requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
mysqlclient>=2.1.0
